+++ xapian-core/backends/glass/glass_check.cc
@@ -155,12 +155,12 @@ GlassTableCheck::block_check(Glass::Cursor * C_, int j, int opts,
 {
     const byte * p = C_[j].get_p();
     uint4 n = C_[j].get_n();
-    size_t c;
-    size_t significant_c = j == 0 ? DIR_START : DIR_START + D2;
+    int c;
+    int significant_c = j == 0 ? DIR_START : DIR_START + D2;
 	/* the first key in an index block is dummy, remember */
 
     size_t max_free = MAX_FREE(p);
-    size_t dir_end = DIR_END(p);
+    int dir_end = DIR_END(p);
     int total_free = block_size - dir_end;
 
     if (!flcheck.mark_used(n))
@@ -169,7 +169,7 @@ GlassTableCheck::block_check(Glass::Cursor * C_, int j, int opts,
     if (j != GET_LEVEL(p))
 	failure("wrong level", n);
     // dir_end must be > DIR_START, fit within the block, and be odd.
-    if (dir_end <= DIR_START || dir_end > block_size || (dir_end & 1) != 1)
+    if (dir_end <= DIR_START || dir_end > int(block_size) || (dir_end & 1) != 1)
 	failure("directory end pointer invalid", n);
 
     if (opts & Xapian::DBCHECK_SHORT_TREE)
@@ -184,7 +184,7 @@ GlassTableCheck::block_check(Glass::Cursor * C_, int j, int opts,
 	    int o = item.get_address() - p;
 	    if (o > int(block_size))
 		failure("item starts outside block", n, c);
-	    if (o - dir_end < max_free)
+	    if (o - dir_end < int(max_free))
 		failure("item overlaps directory", n, c);
 
 	    int kt_len = item.size();
@@ -201,7 +201,7 @@ GlassTableCheck::block_check(Glass::Cursor * C_, int j, int opts,
 	    int o = item.get_address() - p;
 	    if (o > int(block_size))
 		failure("item starts outside block", n, c);
-	    if (o - dir_end < max_free)
+	    if (o - dir_end < int(max_free))
 		failure("item overlaps directory", n, c);
 
 	    int kt_len = item.size();
